+++ app/profile/page.tsx
@@ -1,30 +1,13 @@
 import Link from "next/link";
 import { redirect } from "next/navigation";
-import { supabase } from '@/lib/supabase';
-import { cookies } from "next/headers";
-
-
-async function getUserFromSession() {
-  const cookieStore = await cookies();
-  const session_token = cookieStore.get('session')?.value;
-  if (!session_token) {
-    redirect("/signup");
-  }
-  const {data, error} = await supabase.from('sessions').select('username').eq('cookie_id', session_token);
-  if (error) {
-    redirect("/signup");
-  }
-  // const username = data[0].username;
-  // const { data, error } = await supabase.from('users').select('*').eq('username', username);
-  return data[0];
-}
+import { getUserFromSession } from "@/lib/auth";
 
 export default async function ProfilePage() {
-  const user = await getUserFromSession();
-  if(!user) {
-    console.log(user);
-    redirect('/login');
-  }
+ const user = await getUserFromSession();
+    if(!user) {
+      console.log(user);
+      redirect('/login');
+    }
   const userProfile = {
     username: user.username,
   };
